+++ src/main/java/me/myogoo/ae2fct/init/AE2FCTConfig.java
@@ -7,6 +7,6 @@ import net.neoforged.fml.config.ModConfig;
 public class AE2FCTConfig {
     public static void initialize(ModContainer container) {
         container.registerConfig(ModConfig.Type.COMMON, FluidCraftingConfig.COMMON_SPEC, "ae2fct-common.toml");
-        container.registerConfig(ModConfig.Type.CLIENT, FluidCraftingConfig.ClientSpec, "ae2fct-client.toml");
+        container.registerConfig(ModConfig.Type.CLIENT, FluidCraftingConfig.CLIENT_SPEC, "ae2fct-client.toml");
     }
 }